+++ plugins/guests/darwin/cap/rsync.rb
@@ -17,6 +17,10 @@ module VagrantPlugins
         end
 
         def self.rsync_post(machine, opts)
+          if opts.has_key?(:chown) && !opts[:chown]
+            return
+          end
+
           machine.communicate.sudo(
             "find '#{opts[:guestpath]}' '(' ! -user #{opts[:owner]} -or ! -group #{opts[:group]} ')' -print0 | " +
             "xargs -0 -r chown #{opts[:owner]}:#{opts[:group]}")