+++ plugins/guests/linux/cap/rsync.rb
@@ -17,6 +17,10 @@ module VagrantPlugins
         end
 
         def self.rsync_post(machine, opts)
+          if opts.has_key?(:chown) && !opts[:chown]
+            return
+          end
+
           machine.communicate.sudo(
             "find '#{opts[:guestpath]}' " +
             "'!' -type l -a " +