+++ CHANGELOG.md
@@ -6,6 +6,8 @@ IMPROVEMENTS:
 
 BUG FIXES:
 
+  - commands/package: Package with `--base` for VirtualBox doesn't
+      crash. [GH-3827]
   - provisioners/puppet: Fix setting facter vars with Windows
       guests. [GH-3776]
   - guests/freebsd: Properly register the rsync_pre capability