+++ CHANGELOG.md
@@ -11,6 +11,8 @@ BUG FIXES:
   - provisioners/puppet: Fix setting facter vars with Windows
       guests. [GH-3776]
   - guests/freebsd: Properly register the rsync_pre capability
+  - guests/windows: Certain executed provisioners won't leave output
+      and exit status behind. [GH-3729]
 
 ## 1.6.2 (May 12, 2014)