+++ CHANGELOG.md
@@ -15,6 +15,8 @@ BUG FIXES:
   - guests/freebsd: Properly register the rsync_pre capability
   - guests/windows: Certain executed provisioners won't leave output
       and exit status behind. [GH-3729]
+  - synced\_folders/rsync: `rsync__chown` can be set to `false` to
+      disable recursive chown after sync. [GH-3810]
 
 ## 1.6.2 (May 12, 2014)