+++ CHANGELOG.md
@@ -33,6 +33,8 @@ BUG FIXES:
       disable recursive chown after sync. [GH-3810]
   - synced\_folders/rsync: Use a proper msys path if not in
       Cygwin. [GH-3804]
+  - synced\_folders/rsync: Don't append args infinitely, clear out
+      arg list on each run. [GH-3864]
 
 ## 1.6.2 (May 12, 2014)